\docType{data}
\name{Bearded_effort_ice0}
\alias{Bearded_effort_ice0}
\title{Bearded seal counts and effort information for aerial surveys conducted over the Bering Sea, April-May, 2012}
\description{
   Effort data with some extra zero counts included for the southern ice-free area where no surveying actually occurred.  Similar format to \code{Bearded_effort}}
\author{
  Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{data}

