\docType{data}
\name{Data_for_ST_plot}
\alias{Data_for_ST_plot}
\title{Some extra spatial objects for plotting in the Bering Sea}
\description{
   Spatial objects useful for plotting bearded seal example.  This dataset consists of 
   \itemize{
   \item{\code{alaska_dcw}} A SpatialPolygonsDataFrame giving Alaska land masses
   \item{\code{russia_dcw}} A SpatialPolygonsDataFrame giving Russian land masses
   \item{\code{EEZ_Alaska}} A SpatialPolygonsDataFrame giving Exclusive Economic Zone boundary
   \item{\code{EEZ_s}} A SpatialPolygonsDataFrame giving Exclusive Economic Zone boundary
   \item{\code{Shelf_break}} A SpatialPolygonsDataFrame giving 1000m shelf break
   }
}
\author{
  Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{data}

